+++ rocAL/include/augmentations/augmentations_nodes.h
@@ -56,3 +56,4 @@ THE SOFTWARE.
 #include "augmentations/node_nop.h"
 #include "augmentations/node_sequence_rearrange.h"
 #include "augmentations/audio_augmentations/node_preemphasis_filter.h"
+#include "augmentations/audio_augmentations/node_spectrogram.h"