+++ tests/cpp_api/audio_tests/audio_tests.cpp
@@ -35,7 +35,7 @@ THE SOFTWARE.
 
 using namespace std::chrono;
 
-bool verify_output(float *dst_ptr, long int frames, std::string case_name) {
+bool verify_output(float *dst_ptr, long int frames, long int channels, std::string case_name, int max_samples, int max_channels, int buffer_size) {
     bool pass_status = false;
     // read data from golden outputs
     const char *rocal_data_path = std::getenv("ROCAL_DATA_PATH");
@@ -45,35 +45,46 @@ bool verify_output(float *dst_ptr, long int frames, std::string case_name) {
     }
 
     std::string ref_file_path = std::string(rocal_data_path) + "rocal_data/GoldenOutputsTensor/reference_outputs_audio/" + case_name + "_output.bin";
-    long int out_buffer_size = frames;
-    std::vector<float> ref_output(out_buffer_size);
-    std::fstream fin(ref_file_path, std::ios::in | std::ios::binary);
-    if (fin.is_open()) {
-        for (long int i = 0; i < out_buffer_size; i++) {
-            if (!fin.eof()) {
-                fin.read(reinterpret_cast<char *>(ref_output.data()), sizeof(float));
-            } else {
-                std::cout << "\nUnable to read all data from golden outputs\n";
-                return pass_status;
-            }
-        }
-    } else {
-        std::cout << "\nCould not open the reference output. Please check the path specified\n";
-        return pass_status;
+    std::ifstream fin(ref_file_path, std::ios::binary);  // Open the binary file for reading
+
+    if (!fin.is_open()) {
+        std::cout << "Error: Unable to open the input binary file\n";
+        return 0;
     }
 
+    // Get the size of the file
+    fin.seekg(0, std::ios::end);
+    std::streampos fileSize = fin.tellg();
+    fin.seekg(0, std::ios::beg);
+
+    std::size_t numFloats = fileSize / sizeof(float);
+
+    std::vector<float> ref_output(numFloats);
+
+    // Read the floats from the file
+    fin.read(reinterpret_cast<char *>(ref_output.data()), fileSize);
+
+    if (fin.fail()) {
+        std::cout << "Error: Failed to read from the input binary file\n";
+        return 0;
+    }
+
+    fin.close();
+
     int matched_indices = 0;
-    for (int j = 0; j < frames; j++) {
-        float ref_val, out_val;
-        ref_val = ref_output[j];
-        out_val = dst_ptr[j];
-        bool invalid_comparison = ((out_val == 0.0f) && (ref_val != 0.0f));
-        if (!invalid_comparison && abs(out_val - ref_val) < 1e-20)
-            matched_indices += 1;
+    for (int i = 0; i < frames; i++) {
+        for (int j = 0; j < channels; j++) {
+            float ref_val, out_val;
+            ref_val = ref_output[i * channels + j];
+            out_val = dst_ptr[i * max_channels + j];
+            bool invalid_comparison = ((out_val == 0.0f) && (ref_val != 0.0f));
+            if (!invalid_comparison && std::abs(out_val - ref_val) < 1e-20)
+                matched_indices += 1;
+        }
     }
 
     std::cout << std::endl << "Results for Test case: " << std::endl;
-    if ((matched_indices == frames) && matched_indices != 0) {
+    if ((matched_indices == buffer_size) && matched_indices != 0) {
         pass_status = true;
     }
 
@@ -137,7 +148,7 @@ int test(int test_case, const char *path, int qa_mode, int downmix, int gpu) {
         file_list_path = std::string(std::getenv("ROCAL_DATA_PATH")) + "rocal_data/audio/wav_file_list.txt";
     }
 
-    std::cout << ">>>>>>> Running LABEL READER" << std::endl;
+    std::cout << "Running LABEL READER" << std::endl;
     rocalCreateLabelReader(handle, path, file_list_path.c_str());
 
     if (test_case == 0)
@@ -148,7 +159,7 @@ int test(int test_case, const char *path, int qa_mode, int downmix, int gpu) {
         return -1;
     }
 
-    std::string case_name = "";
+    std::string case_name;
     switch (test_case) {
         case 0: {
             case_name = "audio_decoder";
@@ -161,6 +172,13 @@ int test(int test_case, const char *path, int qa_mode, int downmix, int gpu) {
             RocalAudioBorderType preemph_border_type = RocalAudioBorderType::ROCAL_CLAMP;
             RocalFloatParam p_preemph_coeff = rocalCreateFloatParameter(0.97);
             rocalPreEmphasisFilter(handle, decoded_output, true, p_preemph_coeff, preemph_border_type, tensorOutputType);
+
+        } break;
+        case 2: {
+            std::cout << "Running SPECTROGRAM" << std::endl;
+            case_name = "spectrogram";
+            std::vector<float> window_fn;
+            rocalSpectrogram(handle, decoded_output, true, window_fn, true, true, 2, 512, 320, 160, ROCAL_NFT, ROCAL_FP32);
 
         } break;
         default: {
@@ -178,10 +196,13 @@ int test(int test_case, const char *path, int qa_mode, int downmix, int gpu) {
 
     int iteration = 0;
     float *buffer = nullptr;
-    int frames = 0;
+    int frames = 0, channels = 0;
+    int max_channels = 0;
+    int max_samples = 0;
+    int buffer_size = 0;
     high_resolution_clock::time_point t1 = high_resolution_clock::now();
     while (rocalGetRemainingImages(handle) >= static_cast<size_t>(input_batch_size)) {
-        std::cerr << "\n Iteration:: " << iteration << "\n";
+        std::cout << "\n Iteration:: " << iteration << "\n";
         iteration++;
         if (rocalRun(handle) != 0) {
             break;
@@ -194,18 +215,22 @@ int test(int test_case, const char *path, int qa_mode, int downmix, int gpu) {
         rocalGetImageName(handle, audio_file_name);
         RocalTensorList labels = rocalGetImageLabels(handle);
         int *label_id = reinterpret_cast<int *>(labels->at(0)->buffer());  // The labels are present contiguously in memory
-        std::cerr << "Audio file : " << audio_file_name << "\n";
-        std::cerr << "Label : " << *label_id << "\n";
+        std::cout << "Audio file : " << audio_file_name << "\n";
+        std::cout << "Label : " << *label_id << "\n";
         for (uint idx = 0; idx < output_tensor_list->size(); idx++) {
             buffer = static_cast<float *>(output_tensor_list->at(idx)->buffer());
             output_tensor_list->at(idx)->copy_roi(roi.data());
+            max_channels = output_tensor_list->at(idx)->dims().at(2);
+            max_samples = max_channels == 1 ? 1 : output_tensor_list->at(idx)->dims().at(1);
             frames = roi[idx * 4 + 2];
+            channels = roi[idx * 4 + 3];
+            buffer_size = roi[idx * 4 + 2] * roi[idx * 4 + 3];
         }
     }
 
     if (qa_mode) {
         std::cout << "\n *****************************Verifying Audio output**********************************\n";
-        if (verify_output(buffer, frames, case_name)) {
+        if (verify_output(buffer, frames, channels, case_name, max_samples, max_channels, buffer_size)) {
             std::cout << "PASSED!\n\n";
         } else {
             std::cout << "FAILED!\n\n";
@@ -219,7 +244,7 @@ int test(int test_case, const char *path, int qa_mode, int downmix, int gpu) {
     std::cout << "Decode   time " << rocal_timing.decode_time << std::endl;
     std::cout << "Process  time " << rocal_timing.process_time << std::endl;
     std::cout << "Transfer time " << rocal_timing.transfer_time << std::endl;
-    std::cout << ">>>>> Total Elapsed Time " << dur / 1000000 << " sec " << dur % 1000000 << " us " << std::endl;
+    std::cout << "Total Elapsed Time " << dur / 1000000 << " sec " << dur % 1000000 << " us " << std::endl;
     rocalRelease(handle);
     return 0;
 }