+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * rocDecode for HW decode
 * Support for Audio augmentation - PreEmphasis filter
 * Support for reading from file lists in file reader
+* Support for Audio augmentation - Spectrogram
 
 ### Optimizations